Prion: Refactored prion-toolkit to only contain things that should be different from toolkit
