--- ddp-workspace/projects/prion-toolkit/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/prion-toolkit",
-  "lib": {
-    "entryFile": "src/public-api.ts"
-  }
-}
