--- ddp-workspace/projects/prion-toolkit/src/lib/components/app/app.component.state.ts
@@ -1,3 +1,0 @@
-export interface AppState {
-    joinDialogOpen: boolean;
-}
