--- ddp-workspace/projects/prion-toolkit/src/lib/models/actions/mailingListWorkflowAction.ts
@@ -1,6 +1,0 @@
-import { WorkflowAction } from './workflowAction';
-import { WorkflowActionType } from './../workflowActionType';
-
-export class MailingListWorkflowAction implements WorkflowAction {
-    public actionType = WorkflowActionType.MAILING_LIST;
-}
