--- ddp-workspace/projects/prion-toolkit/src/lib/models/actions/registrationWorkflowAction.ts
@@ -1,6 +1,0 @@
-import { WorkflowAction } from './workflowAction';
-import { WorkflowActionType } from '../workflowActionType';
-
-export class RegistrationWorkflowAction implements WorkflowAction {
-    public actionType = WorkflowActionType.REGISTRATION;
-}
