--- ddp-workspace/projects/prion-toolkit/src/lib/models/actions/urlWorkflowAction.ts
@@ -1,9 +1,0 @@
-import { WorkflowAction } from './workflowAction';
-import { WorkflowActionType } from './../workflowActionType';
-
-export class UrlWorkflowAction implements WorkflowAction {
-    public actionType = WorkflowActionType.URL;
-
-    constructor(
-        public url: string) { }
-}
