--- ddp-workspace/projects/prion-toolkit/src/lib/models/actions/workflowAction.ts
@@ -1,5 +1,0 @@
-import { WorkflowActionType } from '../workflowActionType';
-
-export interface WorkflowAction {
-    actionType: WorkflowActionType;
-}
