--- ddp-workspace/projects/prion-toolkit/src/lib/models/announcementDashboardMessage.ts
@@ -1,5 +1,0 @@
-import { AnnouncementMessage } from 'ddp-sdk';
-
-export interface AnnouncementDashboardMessage extends AnnouncementMessage {
-    shown: boolean;
-}
