--- ddp-workspace/projects/prion-toolkit/src/lib/models/workflowActionType.ts
@@ -1,5 +1,0 @@
-export enum WorkflowActionType {
-    URL = 'URL',
-    MAILING_LIST = 'MAILING_LIST',
-    REGISTRATION = 'REGISTRATION'
-}
