--- ddp-workspace/projects/prion-toolkit/src/lib/models/workflowRule.ts
@@ -1,7 +1,0 @@
-import { WorkflowCommand } from './workflowCommand';
-import { WorkflowAction } from './../models/actions/workflowAction';
-
-export interface WorkflowRule {
-    type: string;
-    func: (action: WorkflowAction) => WorkflowCommand;
-}
