--- ddp-workspace/projects/prion-toolkit/src/lib/models/workflowState.ts
@@ -1,12 +1,0 @@
-export enum WorkflowState {
-    ACTIVITY = 'ACTIVITY',
-    DASHBOARD = 'DASHBOARD',
-    DONE = 'DONE',
-    RETURN_USER = 'RETURN_USER',
-    START = 'START',
-    THANK_YOU = 'THANK_YOU',
-    INTERNATIONAL_PATIENTS = 'INTERNATIONAL_PATIENTS',
-    MAILING_LIST = 'MAILING_LIST',
-    UNKNOWN = 'UNKNOWN',
-    REGISTRATION = 'REGISTRATION'
-}
