--- ddp-workspace/projects/prion-toolkit/src/lib/services/commands/registrationCommand.service.ts
@@ -1,12 +1,0 @@
-import { WorkflowCommand } from '../../models/workflowCommand';
-import { Auth0AdapterService } from 'ddp-sdk';
-import { Observable } from 'rxjs';
-
-export class RegistrationCommand implements WorkflowCommand {
-    constructor(private auth0: Auth0AdapterService) { }
-
-    public execute(): Observable<void> {
-        this.auth0.signup();
-        return new Observable(e => e);
-    }
-}
