--- ddp-workspace/projects/prion-toolkit/src/lib/services/commands/urlCommand.service.ts
@@ -1,15 +1,0 @@
-import { Router } from '@angular/router';
-import { Observable } from 'rxjs';
-import { WorkflowCommand } from '../../models/workflowCommand';
-
-export class UrlCommand implements WorkflowCommand {
-    constructor(
-        private url: string,
-        private router: Router) { }
-
-    public execute(): Observable<void> {
-        this.router.navigateByUrl(this.url);
-        // We have navigated away so there's nothing emitted on this observable.
-        return new Observable(e => e);
-    }
-}
