--- ddp-workspace/projects/prion-toolkit/src/public-api.ts
@@ -1,33 +1,0 @@
-/*
- * Public API Surface of prion-toolkit
- */
-
-export * from './lib/prion-toolkit.module';
-
-export * from './lib/services/communication.service';
-export * from './lib/services/toolkitConfiguration.service';
-export * from './lib/services/workflowBuilder.service';
-export * from './lib/services/workflowMapper.service';
-
-export * from './lib/components/activity/activity.component';
-export * from './lib/components/activity-page/activityPage.component';
-export * from './lib/components/app/app.component';
-export * from './lib/components/common-landing/commonLanding.component';
-export * from './lib/components/workflow-start-activity/workflowStartActivity.component';
-export * from './lib/components/dashboard/dashboard.component';
-export * from './lib/components/dialogs/disclaimer.component';
-export * from './lib/components/dialogs/resendEmail.component';
-export * from './lib/components/dialogs/sessionWillExpire.component';
-export * from './lib/components/dialogs/warning.component';
-export * from './lib/components/error/error.component';
-export * from './lib/components/footer/footer.component';
-export * from './lib/components/header/header.component';
-export * from './lib/components/login-landing/loginLanding.component';
-export * from './lib/components/password/password.component';
-export * from './lib/components/redirect-to-login-landing/redirectToLoginLanding.component';
-export * from './lib/components/warning-message/warning-message.component';
-export * from './lib/components/activity/activityLink.component';
-export * from './lib/components/session-expired/sessionExpired.component';
-export * from './lib/components/redirect-to-auth0-login/redirectToAuth0Login.component';
-
-export * from './lib/guards/headerAction.guard';
