--- ddp-workspace/projects/prion-toolkit/tslint.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tslint.json",
-  "rules": {
-    "directive-selector": [
-      false,
-      "attribute",
-      "lib",
-      "camelCase"
-    ],
-    "component-selector": [
-      false,
-      "element",
-      "prion-toolkit",
-      "kebab-case"
-    ]
-  }
-}
