+++ ddp-workspace/projects/toolkit-prion/src/lib/services/toolkitConfiguration.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+import { ToolkitConfigurationService } from "toolkit";
+
+@Injectable()
+export class PrionToolkitConfigurationService extends ToolkitConfigurationService {
+  assetsBucketUrl: string;
+}
